Organized login provider types into components
